fix warn: - Reports leading whitespaces before Go directives in comments. - Incorrect usages of fmt.Printf, fmt.Println, and similar formatting and printing functions. - Redundant parentheses in expressions and types. - Types in variable and constant declarations that can be omitted since they can be inferred by the compiler. Such types are redundant, omitting them may improve readability of the code. - Type conversions that may be omitted. - Aliases of imported packages that may be omitted.

+++ common/policies/policy.go
@@ -214,7 +214,7 @@ func NewManagerImpl(path string, providers map[int32]Provider, root *cb.ConfigGr
 			}
 			cPolicy = imp
 		} else {
-			provider, ok := providers[int32(policy.Type)]
+			provider, ok := providers[policy.Type]
 			if !ok {
 				return nil, fmt.Errorf("policy %s at path %s has unknown policy type: %v", policyName, path, policy.Type)
 			}
